Debug deployment

+++ app/lib/patient-storage.ts
@@ -3,17 +3,19 @@
 import { Patient } from "../patient-intake/types";
 
 export function savePatientIntake(value: Patient) {
+    if (typeof window !== "undefined") {
+        const localData = localStorage.getItem('patients');
+        const parsedData = localData ? JSON.parse(localData) : [];
 
-    const localData = localStorage.getItem('patients');
+        const phoneExists = parsedData.some((user: Patient) => user.phone === value.phone);
 
-    const parsedData = localData ? JSON.parse(localData) : [];
-
-    const phoneExists = parsedData.some((user:Patient) => user.phone === value.phone);
-
-    phoneExists? null :parsedData.push(value);
-
-    localStorage.setItem('patients', JSON.stringify(parsedData))
+        if (!phoneExists) {
+            parsedData.push(value);
+            localStorage.setItem('patients', JSON.stringify(parsedData));
+        }
+    }
 }
+
 
 export function PatientFetch(){
     const data = localStorage.getItem('patients');
